+++ app/Http/Controllers/backend/SingleSendController.php
@@ -7,7 +7,6 @@ use App\Http\Controllers\Controller;
 use App\Models\Clist;
 use App\Models\SenderVerification;
 use App\Models\SingleSend;
-use App\Models\SuppressionGroup;
 use App\Models\UnsubscribeGroup;
 use Illuminate\Http\Request;
 
@@ -35,7 +34,7 @@ class SingleSendController extends Controller
     /**
      * Show the form for creating a new resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function create()
     {
@@ -49,47 +48,65 @@ class SingleSendController extends Controller
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function store(Request $request)
     {
-        $this->singleSend->addSingleSend($request);
-        return redirect()->route('dashboard.single-sends.index')->with('success','Single send Listed');
+
+        $response = $this->singleSend->addSingleSend($request);
+        return $response;
+        if ($response == 1)
+        {
+            return redirect()->route('dashboard.single-sends.index')->with('success','Single send Listed');
+        }else{
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Something is happened! with sendgrid configuration');
+        }
+
     }
 
 
     public function updateSchedule(Request $request)
     {
-        $this->singleSend->scheduleSingleSends($request->singleSendID);
-        return redirect()->route('dashboard.single-sends.index')->with('success','Single send scheduled for sent after 5 minutes');
+        $response = $this->singleSend->scheduleSingleSends($request->singleSendID);
+        if ($response == 1)
+        {
+            return redirect()->route('dashboard.single-sends.index')->with('success','Single send scheduled for sent after 5 minutes');
+        }else{
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Something is happened! with sendgrid configuration');
+        }
     }
 
 
     public function cancelSchedule(Request $request)
     {
-        $this->singleSend->unscheduledSingleSends($request->singleSendID);
-        return redirect()->route('dashboard.single-sends.index')->with('danger','Single send schedule is cancel');
+        $response = $this->singleSend->unscheduledSingleSends($request->singleSendID);
+        if ($response == 1)
+        {
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Single send schedule is cancel');
+        }else{
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Something is happened! with sendgrid configuration');
+        }
+
     }
 
     /**
      * Display the specified resource.
      *
      * @param  \App\Models\SingleSend  $single_send
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function show(SingleSend $single_send)
     {
-        $list = Clist::where('sendgrid_id', $single_send->list_ids)->firstOrFail();
         $suppression = UnsubscribeGroup::where('sendgrid_id', $single_send->suppression_group_id)->firstOrFail();
         $sender = SenderVerification::where('sendgrid_id', $single_send->sender_id)->firstOrFail();
-        return view('layouts.backend.singleSend.view',compact('single_send','list','suppression','sender'));
+        return view('layouts.backend.singleSend.view',compact('single_send','suppression','sender'));
     }
 
     /**
      * Display the specified resource.
      *
      * @param  \App\Models\SingleSend  $mailId
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function viewMail(SingleSend $mailId)
     {
@@ -100,13 +117,10 @@ class SingleSendController extends Controller
      * Show the form for editing the specified resource.
      *
      * @param  \App\Models\SingleSend  $single_send
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function edit(SingleSend $single_send)
     {
-//        return $single_send;
-//        return view('layouts.backend.singleSend.edit',compact('single_send'));
-
         $contactList = Clist::latest()->get();
         $suppression_group_id = UnsubscribeGroup::latest()->get();
         $sender = SenderVerification::latest()->get();
@@ -119,25 +133,36 @@ class SingleSendController extends Controller
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \App\Models\SingleSend  $single_send
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request,SingleSend $single_send)
     {
-        //return $single_send;
         $this->singleSend->updateSingleSend($request , $single_send);
-        return redirect()->route('dashboard.single-sends.index')->with('success','Single Send Updated');
+        if ($response == 1)
+        {
+            return redirect()->route('dashboard.single-sends.index')->with('success','Single Send Updated');
+        }else{
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Something is happened! with sendgrid configuration');
+        }
+
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  \App\Models\SingleSend  $single_send
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\RedirectResponse
      */
     public function destroy(SingleSend $single_send)
     {
-        $this->singleSend->deleteSingleSend($single_send->sendgrid_id);
-        $single_send->delete();
-        return redirect()->route('dashboard.single-sends.index')->with('danger','Single send Deleted');
+        $response = $this->singleSend->deleteSingleSend($single_send->sendgrid_id);
+        if ($response == 1)
+        {
+            $single_send->lists()->detach();
+            $single_send->delete();
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Single send Deleted');
+        }else{
+            return redirect()->route('dashboard.single-sends.index')->with('danger','Something is happened! with sendgrid configuration');
+        }
     }
 }
